requirements.txt
# IMU Racket Visualization Dependencies
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# 3D Visualization
vpython>=7.6.0

# Serial communication
pyserial>=3.5

# Optional: For enhanced data analysis (if needed)
pandas>=1.3.0
matplotlib>=3.5.0 